requirements.txt
numpy~=1.23.5
vk~=3.0
Pillow~=9.0.1
requests~=2.25.1
joblib~=1.2.0
dependency_injector~=4.41.0
pyTelegramBotAPI~=4.10.0